requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv

sqlalchemy
alembic
asyncpg
psycopg2-binary

python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-multipart
email-validator
authlib

httpx

langchain
langchain-openai
langchain-google-genai
langgraph
openai
chromadb

pandas
openpyxl

google-api-python-client
google-auth
google-auth-oauthlib

pytest
pytest-asyncio
ruff
